requirements.txt
python-telegram-bot[job-queue]==20.7
psycopg2-binary==2.9.9 